+++ app/docspace/people/people.js
@@ -1,10 +1,14 @@
 //
-// (c) Copyright Ascensio System SIA 2024
+// (c) Copyright Ascensio System SIA 2025
 //
 
 // @ts-check
 
 const { Service } = require("../client/client")
+
+/**
+ * @typedef {import("../client/client.js").Filters} Filters
+ */
 
 /**
  * @typedef {Object} Account
@@ -78,6 +82,18 @@ class PeopleService extends Service {
   inviteUser(body) {
     const url = this.client.url("/people/invite")
     return this.client.request("POST", url, body)
+  }
+
+  /**
+   * ```http
+   * GET /people/filter
+   * ```
+   * @param {Filters} filters
+   * @returns {Promise<User[]>}
+   */
+  listGuests(filters) {
+    const url = this.client.url("/people/filter", filters)
+    return this.client.request("GET", url)
   }
 
   /**