+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.1.0
+- new inviteGuest action
+- updated room type
+- updated role list for inviting users
+- available invitation roles now depend on the current user's role
+
 ## 1.0.1
 - fix security vulnerability with z.request() and z.stashFile()